requirements.txt
pypandoc_binary
pypdf
